# Performance improvements for WebGL shape drawing